requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
blinker==1.8.2
bokeh==3.4.1
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.1
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
docopt==0.6.2
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.110.3
fastjsonschema==2.20.0
Flask==3.0.3
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huey==2.5.1
idna==3.7
importlib_resources==6.4.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kombu==5.3.7
langchain==0.2.7
langchain-cli==0.0.25
langchain-community==0.2.7
langchain-core==0.2.18
langchain-experimental==0.0.62
langchain-groq==0.1.6
langchain-openai==0.1.16
langchain-text-splitters==0.2.1
langchainhub==0.1.20
langgraph==0.1.5
langserve==0.2.2
langsmith==0.1.81
libcst==1.4.0
limits==3.13.0
linkify-it-py==2.0.3
llama-index==0.10.27
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.46
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.15
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.25
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.4
llamaindex-py-client==0.1.19
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
mistune==3.0.2
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.35.3
ordered-set==4.1.0
orjson==3.10.5
packaging==24.0
panda==0.3.1
pandas==2.2.2
pandocfilters==1.5.1
panel==1.4.4
param==2.1.0
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
pipreq==0.4
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
pure_eval==0.2.3
pyarrow==15.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
Pygments==2.17.2
pylance==0.14.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyparsing==3.1.2
pypdf==4.2.0
pyproject-toml==0.0.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
pyviz_comms==3.0.2

PyYAML==6.0.1
pyzmq==26.2.0
redis==5.0.8
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rq==1.16.2
rsa==4.9
schedule==1.2.2
setuptools==70.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.37.2
streamlit==1.34.0
striprtf==0.0.26
tabulate==0.9.0
tavily-python==0.3.3
tenacity==8.3.0
tiktoken==0.7.0
tinycss2==1.3.0
toml==0.10.2
tomlkit==0.12.5
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typer==0.9.4
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.2.1
utils==1.0.2
uvicorn==0.23.2
vine==5.1.0
watchdog==4.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
WTForms==3.1.2
xyzservices==2024.6.0
yarg==0.1.9
yarl==1.9.4
